refactor(graphs): refactor features and step-defs to use with new pbraiders library
